% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulasi Data dari Berbagai Distribusi Statistik}
\usage{
simulate_data(n = 100, dist = "normal", params = list())
}
\arguments{
\item{n}{Integer. Jumlah data yang ingin disimulasikan.}

\item{dist}{String. Nama distribusi yang ingin digunakan.}

\item{params}{List. Parameter yang diperlukan untuk distribusi tersebut.}
}
\value{
Vektor numerik berisi data simulasi dari distribusi yang dipilih.
}
\description{
Fungsi ini digunakan untuk mensimulasikan data dari berbagai distribusi statistik,
baik diskrit maupun kontinu. Distribusi yang didukung antara lain:
\itemize{
  \item \strong{"normal"}: distribusi Normal (mean, sd)
  \item \strong{"exponential"}: distribusi Eksponensial (rate)
  \item \strong{"gamma"}: distribusi Gamma (shape, rate)
  \item \strong{"uniform"}: distribusi Uniform (min, max)
  \item \strong{"lognormal"}: distribusi Log-normal (meanlog, sdlog)
  \item \strong{"binomial"}: distribusi Binomial (size, prob)
  \item \strong{"poisson"}: distribusi Poisson (lambda)
  \item \strong{"geometric"}: distribusi Geometrik (prob)
}
}
\examples{
simulate_data(100, "normal", list(mean = 0, sd = 1))
simulate_data(100, "poisson", list(lambda = 3))
simulate_data(100, "uniform", list(min = 0, max = 10))

}
