% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_gof.R
\name{visual_gof}
\alias{visual_gof}
\title{Visualisasi Distribusi Data vs Distribusi Teoritis}
\usage{
visual_gof(data, dist, params)
}
\arguments{
\item{data}{Vektor numerik. Data hasil simulasi.}

\item{dist}{String. Nama distribusi yang digunakan.}

\item{params}{List. Parameter dari distribusi yang bersangkutan.}
}
\value{
Tidak mengembalikan objek. Fungsi menghasilkan grafik (histogram + PDF teoritis).
}
\description{
Fungsi ini menampilkan histogram dari data yang disimulasikan, lalu menambahkan kurva fungsi kepadatan probabilitas (PDF) dari distribusi teoritis sebagai pembanding. Fungsi ini hanya berlaku untuk distribusi kontinu seperti:
\itemize{
  \item \strong{"normal"}: dnorm
  \item \strong{"exponential"}: dexp
  \item \strong{"gamma"}: dgamma
  \item \strong{"uniform"}: dunif
  \item \strong{"lognormal"}: dlnorm
}
}
\examples{
data <- simulate_data(200, "normal", list(mean = 0, sd = 1))
visual_gof(data, "normal", list(mean = 0, sd = 1))

}
